Update of index, update counter for Application Created/Deleted
