+++ functions/package.json
@@ -15,7 +15,7 @@
   },
   "main": "index.js",
   "dependencies": {
-    "firebase-admin": "^12.6.0",
+    "firebase-admin": "^12.7.0",
     "firebase-functions": "^6.0.1"
   },
   "devDependencies": {
